requirements.txt
# AI Debate Platform - Dependencies

# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1

# LLM API Clients
google-genai==0.3.0  # Gemini API (new package)
anthropic==0.25.0            # Claude API
openai==1.12.0               # For Perplexity (uses OpenAI-compatible API)
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Data handling
pandas==2.1.4
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1

# Logging
loguru==0.7.2

# File handling
pathlib2==2.3.7
jsonschema==4.20.0

# UUID and datetime
python-dateutil==2.8.2
